+++ tchannel/zipkin/tracers.py
@@ -40,12 +40,16 @@ from collections import defaultdict
 
 from .formatters import json_formatter
 from .formatters import thrift_formatter
+from .formatters import i64_to_base64
 from .thrift import TCollector
 from .thrift import constants
+from ..thrift import client_for
 
 log = logging.getLogger('zipkin_tracing')
 
 zipkin_log = logging.getLogger('zipkin')
+
+TCollectorClient = client_for('tcollector', TCollector)
 
 
 class EndAnnotationTracer(object):
@@ -135,23 +139,28 @@ class TChannelZipkinTracer(object):
             A tchannel instance to send the trace info to zipkin server
         """
 
-        from ..thrift import client_for
-
-        # TODO add different thrift service name
-        TCollectorClient = client_for('tcollector', TCollector)
-
         self._tchannel = tchannel
-        self.client = TCollectorClient(self._tchannel)
 
     def record(self, traces):
+
         def submit_callback(f):
             if f.exception():
-                log.error('Fail to submit zipkin trace',
-                          exc_info=f.exc_info())
+                log.error(
+                    'Fail to submit zipkin trace',
+                    exc_info=f.exc_info()
+                )
 
+        fus = []
         for (trace, annotations) in traces:
-            f = self.client.submit(thrift_formatter(trace, annotations))
+            client = TCollectorClient(
+                self._tchannel,
+                protocol_headers={'shardKey': i64_to_base64(trace.trace_id)}
+            )
+            f = client.submit(thrift_formatter(trace, annotations))
             f.add_done_callback(submit_callback)
+            fus.append(f)
+
+        return fus
 
 
 class ZipkinTracer(object):